Title: 두 수의 거듭제곱, Time: 92ms, Memory: 8MB, Status: Passed - CodeTree

+++ 231216/두 수의 거듭제곱/two-numbers-of-squares.java
@@ -9,6 +9,6 @@ public class Main {
         int a = Integer.parseInt(stk.nextToken());
         int b = Integer.parseInt(stk.nextToken());
 
-        System.out.println(Math.pow(a, b));
+        System.out.println((int)Math.pow(a, b));
     }
 }
